requirements.txt
Flask
scikit-learn
matplotlib
pandas
numpy
python-docx
sentence-transformers
torch


pip install -r requirements.txt
